assignment_01/finance-application/workspace/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
channels==3.0.4
Pillow==8.2.0
